+++ src/standards/EthRequireLinear.sol
@@ -20,8 +20,8 @@ import {
  * Eth Require with Linear Curve Intent Standard
  * @dev data
  *   [bytes32] standard - the intent standard identifier
- *   [uint48]  startTime - start time of the curve (in seconds)
- *   [uint24]  deltaTime - amount of time from start until curve caps (in seconds)
+ *   [uint40]  startTime - start time of the curve (in seconds)
+ *   [uint32]  deltaTime - amount of time from start until curve caps (in seconds)
  *   [uint96]  startAmount - starting amount
  *   [uint8]   startAmountMult - starting amount multiplier (final_amount = amount << amountMult)
  *   [uint64]  deltaAmount - amount of change after each second
@@ -96,8 +96,8 @@ contract EthRequireLinear is IIntentStandard {
      */
     function encodeData(
         bytes32 standardId,
-        uint48 startTime,
-        uint24 deltaTime,
+        uint40 startTime,
+        uint32 deltaTime,
         int256 startAmount,
         int256 deltaAmount,
         bool isRelative
